fix: index 수정
